refactor: 각 도메인 Controller에 맞게 CommandLineController 수정 및 패키지 이동
