+++ src/main/java/org/promgrammers/springbootbasic/VoucherManagementApplication.java
@@ -1,6 +1,5 @@
 package org.promgrammers.springbootbasic;
 
-import org.promgrammers.springbootbasic.controller.CommandLineController;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
