--- src/test/java/org/promgrammers/springbootbasic/SpringbootBasicApplicationTests.java
@@ -1,13 +1,0 @@
-package org.promgrammers.springbootbasic;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SpringbootBasicApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
