+++ src/test/java/org/promgrammers/springbootbasic/domain/voucher/repository/JdbcVoucherRepositoryTest.java
@@ -3,7 +3,7 @@ package org.promgrammers.springbootbasic.domain.voucher.repository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.promgrammers.springbootbasic.controller.CommandLineController;
+import org.promgrammers.springbootbasic.CommandLineController;
 import org.promgrammers.springbootbasic.domain.voucher.model.FixedAmountVoucher;
 import org.promgrammers.springbootbasic.domain.voucher.model.PercentDiscountVoucher;
 import org.promgrammers.springbootbasic.domain.voucher.model.Voucher;
